Fix rects by using idividual x, y, width and height.
Using a BRect to store the coordinates and adding the width and height
did make the logic depend on the order of the attributes. When the
width or height was specified before the x or y coordinates
respectively they would be added to a yet empty rect, messing up the
eventual rect.